Misc cleanup & return randomized hashing
