+++ src/mscorlib/src/System/Globalization/CultureInfo.Unix.cs
@@ -6,11 +6,6 @@ namespace System.Globalization
 {
     public partial class CultureInfo : IFormatProvider
     {
-        private static CultureInfo GetUserDefaultCultureCacheOverride()
-        {
-            return null; // ICU doesn't provide a user override
-        }
-
         internal static CultureInfo GetUserDefaultCulture()
         {
             CultureInfo cultureInfo = null;
@@ -32,5 +27,91 @@ namespace System.Globalization
         {
             return GetUserDefaultCulture();
         }
+
+        ////////////////////////////////////////////////////////////////////////
+        //
+        //  CurrentCulture
+        //
+        //  This instance provides methods based on the current user settings.
+        //  These settings are volatile and may change over the lifetime of the
+        //  thread.
+        //
+        ////////////////////////////////////////////////////////////////////////
+
+        //
+        // We use the following order to return CurrentCulture and CurrentUICulture
+        //      o   use current thread culture if the user already set one using CurrentCulture/CurrentUICulture
+        //      o   use thread culture if the user already set one using DefaultThreadCurrentCulture
+        //          or DefaultThreadCurrentUICulture
+        //      o   Use NLS default user culture
+        //      o   Use NLS default system culture
+        //      o   Use Invariant culture
+        //
+        public static CultureInfo CurrentCulture
+        {
+            get
+            {
+                if (Thread.m_CurrentCulture != null)
+                {
+                    return Thread.m_CurrentCulture;
+                }
+
+                ci = s_DefaultThreadCurrentCulture;
+                if (ci != null)
+                {
+                    return ci;
+                }
+
+                // if s_userDefaultCulture == null means CultureInfo statics didn't get initialized yet. this can happen if there early static 
+                // method get executed which eventually hit the cultureInfo code while CultureInfo statics didn’t get chance to initialize
+                if (s_userDefaultCulture == null)
+                {
+                    Init();
+                }
+
+                Debug.Assert(s_userDefaultCulture != null);
+                return s_userDefaultCulture;
+            }
+
+            set
+            {
+                if (value == null)
+                {
+                    throw new ArgumentNullException(nameof(value));
+                }
+                
+                if (s_asyncLocalCurrentCulture == null)
+                {
+                    Interlocked.CompareExchange(ref s_asyncLocalCurrentCulture, new AsyncLocal<CultureInfo>(AsyncLocalSetCurrentCulture), null);
+                }
+                s_asyncLocalCurrentCulture.Value = value;
+            }
+        }
+
+        public static CultureInfo CurrentUICulture
+        {
+            get
+            {
+                return GetCurrentUICultureNoAppX();
+            }
+
+            set
+            {
+                if (value == null)
+                {
+                    throw new ArgumentNullException(nameof(value));
+                }
+
+                CultureInfo.VerifyCultureName(value, true);
+                if (s_asyncLocalCurrentUICulture == null)
+                {
+                    Interlocked.CompareExchange(ref s_asyncLocalCurrentUICulture, new AsyncLocal<CultureInfo>(AsyncLocalSetCurrentUICulture), null);
+                }
+
+                // this one will set s_currentThreadUICulture too
+                s_asyncLocalCurrentUICulture.Value = value;               
+            }
+        }
+        
     }
 }
