+++ src/vm/mscorlib.h
@@ -304,79 +304,6 @@ DEFINE_FIELD_U(m_encodedEnumType,  CustomAttributeType,            m_enumType)
 DEFINE_FIELD_U(m_encodedArrayType, CustomAttributeType,            m_arrayType)
 DEFINE_FIELD_U(m_padding,          CustomAttributeType,            m_padding)
 
-
-
-
-#ifndef FEATURE_COREFX_GLOBALIZATION
-DEFINE_CLASS_U(Globalization,          CultureData,           CultureDataBaseObject)
-DEFINE_FIELD_U(_sRealName,             CultureDataBaseObject,  _sRealName)
-DEFINE_FIELD_U(_sWindowsName,          CultureDataBaseObject,  _sWindowsName)
-DEFINE_FIELD_U(_sName,                 CultureDataBaseObject,  _sName)
-DEFINE_FIELD_U(_sParent,               CultureDataBaseObject,  _sParent)
-DEFINE_FIELD_U(_sLocalizedDisplayName, CultureDataBaseObject,  _sLocalizedDisplayName)
-DEFINE_FIELD_U(_sEnglishDisplayName,   CultureDataBaseObject,  _sEnglishDisplayName)
-DEFINE_FIELD_U(_sNativeDisplayName,    CultureDataBaseObject,  _sNativeDisplayName)
-DEFINE_FIELD_U(_sSpecificCulture,      CultureDataBaseObject,  _sSpecificCulture)
-DEFINE_FIELD_U(_sISO639Language,       CultureDataBaseObject,  _sISO639Language)
-DEFINE_FIELD_U(_sISO639Language2,       CultureDataBaseObject, _sISO639Language2)
-DEFINE_FIELD_U(_sLocalizedLanguage,    CultureDataBaseObject,  _sLocalizedLanguage)
-DEFINE_FIELD_U(_sEnglishLanguage,      CultureDataBaseObject,  _sEnglishLanguage)
-DEFINE_FIELD_U(_sNativeLanguage,       CultureDataBaseObject,  _sNativeLanguage)
-DEFINE_FIELD_U(_sAbbrevLang,           CultureDataBaseObject,  _sAbbrevLang)
-DEFINE_FIELD_U(_sConsoleFallbackName,  CultureDataBaseObject,  _sConsoleFallbackName)
-DEFINE_FIELD_U(_iInputLanguageHandle,  CultureDataBaseObject,  _iInputLanguageHandle)
-DEFINE_FIELD_U(_sRegionName,           CultureDataBaseObject,  _sRegionName)
-DEFINE_FIELD_U(_sLocalizedCountry,     CultureDataBaseObject,  _sLocalizedCountry)
-DEFINE_FIELD_U(_sEnglishCountry,       CultureDataBaseObject,  _sEnglishCountry)
-DEFINE_FIELD_U(_sNativeCountry,        CultureDataBaseObject,  _sNativeCountry)
-DEFINE_FIELD_U(_sISO3166CountryName,   CultureDataBaseObject,  _sISO3166CountryName)
-DEFINE_FIELD_U(_sISO3166CountryName2,  CultureDataBaseObject,  _sISO3166CountryName2)
-DEFINE_FIELD_U(_iGeoId,                CultureDataBaseObject,  _iGeoId)
-DEFINE_FIELD_U(_sPositiveSign,         CultureDataBaseObject,  _sPositiveSign)
-DEFINE_FIELD_U(_sNegativeSign,         CultureDataBaseObject,  _sNegativeSign)
-DEFINE_FIELD_U(_iDigits,               CultureDataBaseObject,  _iDigits)
-DEFINE_FIELD_U(_iNegativeNumber,       CultureDataBaseObject,  _iNegativeNumber)
-DEFINE_FIELD_U(_waGrouping,            CultureDataBaseObject,  _waGrouping)
-DEFINE_FIELD_U(_sDecimalSeparator,     CultureDataBaseObject,  _sDecimalSeparator)
-DEFINE_FIELD_U(_sThousandSeparator,    CultureDataBaseObject,  _sThousandSeparator)
-DEFINE_FIELD_U(_sNaN,                  CultureDataBaseObject,  _sNaN)
-DEFINE_FIELD_U(_sPositiveInfinity,     CultureDataBaseObject,  _sPositiveInfinity)
-DEFINE_FIELD_U(_sNegativeInfinity,     CultureDataBaseObject,  _sNegativeInfinity)
-DEFINE_FIELD_U(_iNegativePercent,      CultureDataBaseObject,  _iNegativePercent)
-DEFINE_FIELD_U(_iPositivePercent,      CultureDataBaseObject,  _iPositivePercent)
-DEFINE_FIELD_U(_sPercent,              CultureDataBaseObject,  _sPercent)
-DEFINE_FIELD_U(_sPerMille,             CultureDataBaseObject,  _sPerMille)
-DEFINE_FIELD_U(_sCurrency,             CultureDataBaseObject,  _sCurrency)
-DEFINE_FIELD_U(_sIntlMonetarySymbol,   CultureDataBaseObject,  _sIntlMonetarySymbol)
-DEFINE_FIELD_U(_sEnglishCurrency,      CultureDataBaseObject,  _sEnglishCurrency)
-DEFINE_FIELD_U(_sNativeCurrency,       CultureDataBaseObject,  _sNativeCurrency)
-DEFINE_FIELD_U(_iCurrencyDigits,       CultureDataBaseObject,  _iCurrencyDigits)
-DEFINE_FIELD_U(_iCurrency,             CultureDataBaseObject,  _iCurrency)
-DEFINE_FIELD_U(_iNegativeCurrency,     CultureDataBaseObject,  _iNegativeCurrency)
-DEFINE_FIELD_U(_waMonetaryGrouping,    CultureDataBaseObject,  _waMonetaryGrouping)
-DEFINE_FIELD_U(_sMonetaryDecimal,      CultureDataBaseObject,  _sMonetaryDecimal)
-DEFINE_FIELD_U(_sMonetaryThousand,     CultureDataBaseObject,  _sMonetaryThousand)
-DEFINE_FIELD_U(_iMeasure,              CultureDataBaseObject,  _iMeasure)
-DEFINE_FIELD_U(_sListSeparator,        CultureDataBaseObject,  _sListSeparator)
-DEFINE_FIELD_U(_sAM1159,               CultureDataBaseObject,  _sAM1159)
-DEFINE_FIELD_U(_sPM2359,               CultureDataBaseObject,  _sPM2359)
-DEFINE_FIELD_U(_sTimeSeparator,        CultureDataBaseObject,  _sTimeSeparator)
-DEFINE_FIELD_U(_saLongTimes,           CultureDataBaseObject,  _saLongTimes)
-DEFINE_FIELD_U(_saShortTimes,          CultureDataBaseObject,  _saShortTimes)
-DEFINE_FIELD_U(_saDurationFormats,     CultureDataBaseObject,  _saDurationFormats)
-DEFINE_FIELD_U(_iFirstDayOfWeek,       CultureDataBaseObject,  _iFirstDayOfWeek)
-DEFINE_FIELD_U(_iFirstWeekOfYear,      CultureDataBaseObject,  _iFirstWeekOfYear)
-DEFINE_FIELD_U(_waCalendars,           CultureDataBaseObject,  _waCalendars)
-DEFINE_FIELD_U(_calendars,             CultureDataBaseObject,  _calendars)
-DEFINE_FIELD_U(_iReadingLayout,        CultureDataBaseObject,  _iReadingLayout)
-DEFINE_FIELD_U(_iDefaultAnsiCodePage,  CultureDataBaseObject,  _iDefaultAnsiCodePage)
-DEFINE_FIELD_U(_iDefaultOemCodePage,   CultureDataBaseObject,  _iDefaultOemCodePage)
-DEFINE_FIELD_U(_iDefaultMacCodePage,   CultureDataBaseObject,  _iDefaultMacCodePage)
-DEFINE_FIELD_U(_iDefaultEbcdicCodePage,CultureDataBaseObject,  _iDefaultEbcdicCodePage)
-DEFINE_FIELD_U(_iLanguage,             CultureDataBaseObject,  _iLanguage)
-DEFINE_FIELD_U(_bUseOverrides,         CultureDataBaseObject,  _bUseOverrides)
-DEFINE_FIELD_U(_bNeutral,              CultureDataBaseObject,  _bNeutral)
-#endif
 #ifndef FEATURE_COREFX_GLOBALIZATION
 DEFINE_CLASS_U(Globalization,          CalendarData,           CalendarDataBaseObject)
 DEFINE_FIELD_U(sNativeName,            CalendarDataBaseObject, sNativeName)
